One netstandard2.0 library
 - The lowest framework supported is 4.6.2
 - Add CodePages to netstandard project to support net6
 - net6 tests prove that this closes #43
