+++ netstandard20/netstandard.csproj
@@ -1,23 +1,48 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Description>Zipstorer</Description>
+    <Description>A Pure C# class to store files in Zip</Description>
     <AssemblyTitle>ZipStorer</AssemblyTitle>
-    <VersionPrefix>3.8.0</VersionPrefix>
+    <VersionPrefix>$(AssemblyVersion)</VersionPrefix>
     <Authors>Jaime Olivares</Authors>
     <TargetFramework>netstandard2.0</TargetFramework>
     <AssemblyName>ZipStorer</AssemblyName>
     <PackageId>netstandard</PackageId>
-    <NetStandardImplicitPackageVersion>2.0.0</NetStandardImplicitPackageVersion>
     <OutputType>Library</OutputType>
+    <AssemblyVersion>3.9.0</AssemblyVersion>
+    <Version>$(AssemblyVersion)</Version>
+    <FileVersion>$(AssemblyVersion)</FileVersion>
+    <EnableNETAnalyzers>True</EnableNETAnalyzers>
+    <EnforceCodeStyleInBuild>True</EnforceCodeStyleInBuild>
+    <AnalysisLevel>latest-recommended</AnalysisLevel>
+    <ProduceReferenceAssembly>True</ProduceReferenceAssembly>
+    <GenerateDocumentationFile>True</GenerateDocumentationFile>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <Copyright>Copyright © Jaime Olivares 2019</Copyright>
+    <Title>ZipStorer</Title>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <PackageProjectUrl>https://github.com/jaime-olivares/zipstorer</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/jaime-olivares/zipstorer.git</RepositoryUrl>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    <CheckForOverflowUnderflow>True</CheckForOverflowUnderflow>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+    <CheckForOverflowUnderflow>True</CheckForOverflowUnderflow>
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\src\zipstorer.cs" Exclude="bin\**;obj\**;**\*.xproj;packages\**" Link="ZipStorer.cs" />
+    <None Include="..\README.md">
+      <Pack>True</Pack>
+      <PackagePath>\</PackagePath>
+    </None>
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="System.Runtime" Version="*" />
+    <PackageReference Include="System.Text.Encoding.CodePages" Version="7.0.0" />
     <PackageReference Include="System.Threading.Thread" Version="*" />
   </ItemGroup>
