+++ test/UnitTestWrite.cs
@@ -15,8 +15,6 @@ namespace Test
         [ClassInitialize]
         public static void Initialize(TestContext test)
         {
-            Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
-
             string content = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
             buffer = Encoding.UTF8.GetBytes(content);
         }
