+++ zipstorer.sln
@@ -1,14 +1,11 @@
-﻿Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26228.10
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.7.34221.43
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "net20", "net20\net20.csproj", "{69B1B1EA-F10C-4F8E-BA08-48A16B993305}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "net45", "net45\net45.csproj", "{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "netstandard", "netstandard20\netstandard.csproj", "{9784E5E4-C509-40E5-B61F-E08256A669C9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "test", "test\test.csproj", "{06CC1951-28C6-4BD0-8CE5-C4258632B318}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "test", "test\test.csproj", "{06CC1951-28C6-4BD0-8CE5-C4258632B318}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,18 +17,6 @@ Global
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Debug|x64.Build.0 = Debug|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Debug|x86.Build.0 = Debug|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Release|x64.ActiveCfg = Release|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Release|x64.Build.0 = Release|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Release|x86.ActiveCfg = Release|Any CPU
-		{1A30C0E0-46B7-4939-BD6B-1293B197F6B0}.Release|x86.Build.0 = Release|Any CPU
 		{9784E5E4-C509-40E5-B61F-E08256A669C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9784E5E4-C509-40E5-B61F-E08256A669C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9784E5E4-C509-40E5-B61F-E08256A669C9}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -56,20 +41,11 @@ Global
 		{06CC1951-28C6-4BD0-8CE5-C4258632B318}.Release|x64.Build.0 = Release|Any CPU
 		{06CC1951-28C6-4BD0-8CE5-C4258632B318}.Release|x86.ActiveCfg = Release|Any CPU
 		{06CC1951-28C6-4BD0-8CE5-C4258632B318}.Release|x86.Build.0 = Release|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Debug|x64.Build.0 = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Debug|x86.Build.0 = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Release|Any CPU.ActiveCfg = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Release|Any CPU.Build.0 = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Release|x64.ActiveCfg = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Release|x64.Build.0 = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Release|x86.ActiveCfg = Debug|Any CPU
-		{69B1B1EA-F10C-4F8E-BA08-48A16B993305}.Release|x86.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6D616F9F-3E8F-4B19-82B2-A64287879201}
 	EndGlobalSection
 EndGlobal
